Cycle 9: Phase 11.1 Lint/Type Sweep — DEBT-001 resolved
Cleared all in-scope ruff and mypy errors. ruff: 18 → 0 + the
pyproject deprecation warning gone (config migrated from
top-level [tool.ruff] select/ignore/isort to [tool.ruff.lint]
and [tool.ruff.lint.isort]). In-scope mypy: 12 → 0 (the 29
remaining mypy errors are all in modules outside DEBT-001's
scope — captured as DEBT-005..008 for follow-up cycles).

Concrete fixes:
- src/ai/claude.py — 2 B904 raise-from chains
- src/strategy/loader.py — 5 B904 chains
- src/strategy/factory.py — Callable from collections.abc (UP035)
- src/ai/improver.py — coerce frontmatter fields to str at parse
  time so _build_filename receives the right type
- src/trading/live.py — Order import + return-type widening
  cascades through 5 attr-defined errors at 235/244/252/438/445;
  Literal["buy","sell"] for closing_side
- src/trading/paper.py — same Literal fix
- src/backtest/analyzer.py — float() cast for no-any-return
- 6 test files — F401/F841/I001 unused imports/locals removed
- pyproject.toml — types-PyYAML>=6.0 added to dev extras
- scripts/lint.sh — new operator script (ruff check + mypy)

Tests unchanged: 1083 pass, no behaviour change, no new tests
(spec explicitly forbade), zero new # noqa / # type: ignore
suppressions.

Five new debts logged for the remaining mypy clusters
(DEBT-005 ccxt typing, DEBT-006 exchange factory shape drift,
DEBT-007 dashboard Streamlit types, DEBT-008 main.py lambda,
DEBT-009 lint.sh --fix unsafe-for-CI). DEBT-001 moved to
Resolved.

+++ src/backtest/analyzer.py
@@ -226,7 +226,7 @@ class PerformanceAnalyzer:
         if growth <= 0:
             # Ruined account; annualization is meaningless.
             return -100.0
-        return (growth ** (DAYS_PER_YEAR / days) - 1) * 100
+        return float((growth ** (DAYS_PER_YEAR / days) - 1) * 100)
 
     @staticmethod
     def _profit_factor(
